Refactor info experiencia
