+++ dream-lab-frontend/src/pages/CrearExperiencia/CrearExperiencia.jsx
@@ -7,11 +7,10 @@ import Stepper from "@mui/material/Stepper";
 import Step from "@mui/material/Step";
 import StepLabel from "@mui/material/StepLabel";
 import GlassCard from "../../GlobalComponents/GlassCard/GlassCard";
-import { DatePicker, Select, SelectItem } from "@nextui-org/react";
 import SubirImagenBox from "../CrearAnuncioVideowall/components/FormularioCreacionAnuncio/components/SubirImagenBox/SubirImagenBox";
 import AgregarImagen from "src/assets/CrearAnuncioVideowall/agregarImagen.webp";
 import AgregarImagenError from "src/assets/CrearAnuncioVideowall/subirArchivoErroneo.png";
-import { Input } from "@nextui-org/react";
+import InfoExperiencia from "./components/InfoExperiencia";
 
 const steps = [
 	"Crea tu experiencia",
@@ -23,8 +22,6 @@ function CrearExperiencia() {
 	const [page, setPage] = useState(0); // Manage the current page state
 	const [fileSeleccionado, setFileSeleccionado] = useState(null);
 	const [isInvalidFile, setIsInvalidFile] = useState(false);
-	const [caracteresRestantesNom, setCaracteresRestantesNom] = useState(30);
-	const [nombre, setNombre] = useState("");
 
 	function handleSiguiente() {
 		if (page < steps.length - 1) {
@@ -37,11 +34,6 @@ function CrearExperiencia() {
 
 	function handleStepClick(index) {
 		setPage(index); // Move to the clicked step
-	}
-
-	function handleNombreChange() {
-		setNombre(event.target.value);
-		setCaracteresRestantesNom(30 - event.target.value.length);
 	}
 
 	const handleInvalidFileChange = (invalid) => {
@@ -79,33 +71,7 @@ function CrearExperiencia() {
 						{" "}
 						{/* Página 1 */}
 						<div className="portada-tipo-selector-crear-experiencia">
-							<GlassCard classes="dropdowns-container-crear-experiencia">
-								<p className="text-white">Nombre de la experiencia</p>
-								<div className="flex w-full flex-wrap md:flex-nowrap gap-4">
-									<Input
-										type="text"
-										aria-label="Nombre de experiencia"
-										placeholder="Nombre de experiencia"
-										onChange={handleNombreChange}
-										maxLength={30}
-										value={nombre}
-										data-cy="input-nombre-exp"
-									/>
-								</div>
-								<div className="footer-input-formulario-anuncio">
-									({caracteresRestantesNom} caracteres restantes)
-								</div>
-								<p className="text-white">Tipo de experiencia</p>
-								<Select
-									className="mb-3 "
-									aria-label="Tipo de experiencia"
-									placeholder="Selecciona un tipo"
-									data-cy="selector-tipo-exp"
-								>
-									<SelectItem>Autodirigida</SelectItem>
-									<SelectItem>Unidad de Formación</SelectItem>
-								</Select>
-							</GlassCard>
+							<InfoExperiencia/>
 							<GlassCard classes="subir-portada-container-crear-experiencia">
 								<p className="titulo-portada-crear-exp">Portada</p>
 								<SubirImagenBox
